HWGroup/requirements.txt
flask
flask_bootstrap
numpy
matplotlib
gspread
pandas
rpi_ws281x
adafruit-circuitpython-neopixel
Pillow
oauth2client==1.5.2
bitstring